blog_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
